requirements.txt
Click==7.0
Flask==1.0.2
Flask-Login==0.4.1
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
mutagen==1.41.0
mysql-connector==2.2.9
mysql-connector-python==8.0.15
requests==2.21.0
virtualenv==15.1.0
Werkzeug==0.14.1

